Rename simple_bet module to single_bet
