+++ crapssim/strategy/single_bet.py
